% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{build_cache_path}
\alias{build_cache_path}
\title{Build cache file path}
\usage{
build_cache_path(end_year, cache_type)
}
\arguments{
\item{end_year}{School year end}

\item{cache_type}{Type of cache ("tidy" or "wide")}
}
\value{
File path string
}
\description{
Build cache file path
}
\keyword{internal}
