% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{build_entities_url}
\alias{build_entities_url}
\title{Build CAASPP entities file URL}
\usage{
build_entities_url(end_year, format = c("csv", "ascii"))
}
\arguments{
\item{end_year}{School year end}

\item{format}{One of "csv" or "ascii"}
}
\value{
URL string
}
\description{
Build CAASPP entities file URL
}
\keyword{internal}
