% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{check_caaspp_url}
\alias{check_caaspp_url}
\title{Check if CAASPP URL is accessible}
\usage{
check_caaspp_url(url)
}
\arguments{
\item{url}{URL to test}
}
\value{
TRUE if accessible, FALSE otherwise
}
\description{
Tests if a CAASPP research file URL is accessible.
}
\keyword{internal}
