% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_enrollment.R
\name{enr_grade_aggs}
\alias{enr_grade_aggs}
\title{Create grade-level aggregates}
\usage{
enr_grade_aggs(df)
}
\arguments{
\item{df}{A tidy enrollment data frame}
}
\value{
Data frame with aggregated enrollment for grade bands
}
\description{
Creates aggregations for common grade groupings: K-8, 9-12 (HS), K-12.
}
