% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_graduation.R
\name{get_available_grad_years}
\alias{get_available_grad_years}
\title{Get available graduation years}
\usage{
get_available_grad_years()
}
\value{
Integer vector of years (2017-2025)
}
\description{
Returns a vector of years for which graduation rate data is available
from the California Department of Education.
}
\examples{
\dontrun{
get_available_grad_years()
# Returns: 2017 2018 2019 ... 2025
}
}
