% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_raw_enr_modern}
\alias{get_raw_enr_modern}
\title{Download modern Census Day format (2024+) CDE data}
\usage{
get_raw_enr_modern(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Raw data frame
}
\description{
Download modern Census Day format (2024+) CDE data
}
\keyword{internal}
