% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_graduation.R
\name{get_raw_graduation}
\alias{get_raw_graduation}
\title{Download raw graduation data from CDE}
\usage{
get_raw_graduation(end_year)
}
\arguments{
\item{end_year}{School year end (2023-24 = 2024). Valid years: 2017-2025.}
}
\value{
Data frame with graduation data
}
\description{
Downloads graduation rate data Excel file from CDE School Dashboard.
}
\keyword{internal}
