% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{print.ca_assess_data}
\alias{print.ca_assess_data}
\title{Print CAASPP assessment data}
\usage{
\method{print}{ca_assess_data}(x, ...)
}
\arguments{
\item{x}{A ca_assess_data object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print CAASPP assessment data
}
