% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_assessment.R
\name{print.ca_assess_tidy}
\alias{print.ca_assess_tidy}
\title{Print tidy CAASPP assessment data}
\usage{
\method{print}{ca_assess_tidy}(x, ...)
}
\arguments{
\item{x}{A ca_assess_tidy object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print tidy CAASPP assessment data
}
